templates/web-search-app/1.0.0/main.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
联网搜索应用主程序
"""

import yaml
import requests
from typing import Dict, List, Any


class SearchApp:
    """联网搜索应用类"""
    
    def __init__(self, config_path: str = "config.yaml"):
        """初始化搜索应用"""
        self.config = self.load_config(config_path)
    
    def load_config(self, config_path: str) -> Dict:
        """加载配置文件"""
        try:
            with open(config_path, 'r', encoding='utf-8') as f:
                return yaml.safe_load(f)
        except Exception as e:
            raise Exception(f"配置文件加载失败: {e}")
    
    def search(self, query: str, limit: int = 10) -> Dict[str, Any]:
        """执行搜索"""
        try:
            # TODO: 实现具体的搜索逻辑
            results = self._perform_search(query, limit)
            
            return {
                "status": "success",
                "data": results,
                "count": len(results)
            }
        except Exception as e:
            return {
                "status": "error",
                "message": str(e)
            }
    
    def _perform_search(self, query: str, limit: int) -> List[Dict]:
        """执行实际搜索操作"""
        # TODO: 根据配置的搜索引擎实现搜索
        # 这里是示例代码，需要根据实际需求修改
        
        results = []
        # 示例搜索结果
        for i in range(min(limit, 5)):
            results.append({
                "title": f"搜索结果 {i+1}: {query}",
                "url": f"https://example.com/result/{i+1}",
                "snippet": f"这是关于 '{query}' 的搜索结果摘要"
            })
        
        return results


def main():
    """主函数"""
    app = SearchApp()
    
    # 示例搜索
    result = app.search("MaxKB 智能体平台", 5)
    print(result)


if __name__ == "__main__":
    main()